Rerun Swift measurements
I reran the Swift code, using a newly trained network (I realized that the network I was using was
old and no longer compatible with the Swift implementation, which was causing some issues
and causing the triggering early). So I trained a new network and adjusted the audio buffer size.
The net result is longer latency, but less jitter.

In addition, I added tools to count the FP/FN/TP data, although I am not sure we need it
in the paper:

* False negatives: 4
* False positives: 1
* True positives: 632

Because I am just using the single combined audio file, it is hard to separate songs,
so I didn't try to count "true negatives."