Initial pass at Swift runtime
Updated the Swift runtime passage to describe how to install and use the Swift detector. I included instructions for configuring the Arduino, although the Arduino code still has not been merged into the master branch of the Swift detector (but I will finis that before publication). In addition, I added a brief section on customization, pinpointing the relevant file for customizing behavior.